Cambio de Integer a Long por escalabilidad

+++ cliente/cliente/src/main/java/com/taller/cliente/model/Cliente.java
@@ -14,7 +14,7 @@ public class Cliente {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer idCliente;
+    private Long idCliente;
 
     @Column(nullable = false)
     private String prNombre;
